backend/requirements.txt
fastapi
uvicorn
python-multipart
sqlalchemy
psycopg2-binary
python-dotenv
ultralytics
astropy
astroquery
google-generativeai
opencv-python-headless
requests
